+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.4.2] - 2024-01-11
+
+The release incorporates the up-to-date [API contract](openAPIDefinition.json) as of 2024-01-11 05:27:00 GMT.
+
+### Added
+
+- Added support to list the projects shared with the account. Find more [here](https://neon.tech/docs/manage/projects#share-a-project).
+
 ## [v0.4.1] - 2023-12-22
 
 The release incorporates the up-to-date [API contract](openAPIDefinition.json) as of 2023-12-22 01:09:10 GMT.